This PR begins the work on selective execution in hopes of improving performance for du_dl computation. This is needed for faster non-equilibrium calculations which require du_dl on every step. For RBFE calculations with interpolated nonbonded kernels, du_dl is chained via:

```
d/dl u(lambda, params(lambda)) = du/dlambda + du/dparams.dparams/dlambda
```

Currently, with the InterpolatedPotential class, we need to compute a per-particle du/dparams (even though the vast majority of them have ```dparams/dlambda=0```, since the host parameters are not interpolated). However, it is not clear how to pass the ```dparams/dlambda``` information to the underlying Nonbonded class, while still keeping the abstraction barriers clear. So this PR modifies the Nonbonded class to be able to directly do parameter interpolation.

In order to improve performance, we note that we generally require 4 outputs, but not all at once:
```
u (rarely needed)
du/dp (needed at the endpoints)
du/dl (always needed for NEQ, and periodically needed for EQ)
du/dx (always needed)
```
Previously, we had a special ```execute_du_dx_only()``` method to indicate that we only needed ```du_dx```. However, with NEQ support coming down the road, we will also need ```du_dl``` in a way that is performant. I really did not want to add a third ```execute_du_dx_and_du_dl_only()``` class, and specialize a third kernel just for this computation. So with some template magic, we pre-generate all 2^4 choices on the kernel level, and trigger specialized paths depending on which of the outputs we need.

So the kernel code is now templatized with:

```
template <
    typename RealType,
    bool ALCHEMICAL,
    bool COMPUTE_U,
    bool COMPUTE_DU_DX,
    bool COMPUTE_DU_DL,
    bool COMPUTE_DU_DP
>
void __global__ k_nonbonded_unified(...)
```
And specialized code is generated for each combination. The ALCHEMICAL flag is used to partition interacting tiles into two sets, one non-alchemical set where we guarantee that ```du_dl==0``` and ```dp_dl==0``` for every particle in the tile, and one alchemical set where we need to compute the full chain. Note that 95% of the interactions will follow the fast, non-alchemical route. 

Some baseline performance numbers, assuming every tile is alchemical, is as follows:

```
   Avg                                    alchem     u     du_dx   du_dl   du_dp
554.34us  void k_nonbonded_unified<float, bool=1, bool=1, bool=1, bool=1, bool=1>
533.58us  void k_nonbonded_unified<float, bool=1, bool=0, bool=1, bool=1, bool=1>
457.73us  void k_nonbonded_unified<float, bool=1, bool=1, bool=1, bool=1, bool=0>
439.58us  void k_nonbonded_unified<float, bool=1, bool=0, bool=1, bool=1, bool=0> <- du_dx and du_dl
426.43us  void k_nonbonded_unified<float, bool=1, bool=1, bool=1, bool=0, bool=1>
403.66us  void k_nonbonded_unified<float, bool=1, bool=1, bool=0, bool=1, bool=1>
393.92us  void k_nonbonded_unified<float, bool=1, bool=0, bool=1, bool=0, bool=1>
389.28us  void k_nonbonded_unified<float, bool=1, bool=0, bool=0, bool=1, bool=1>
332.31us  void k_nonbonded_unified<float, bool=1, bool=1, bool=1, bool=0, bool=0>
316.21us  void k_nonbonded_unified<float, bool=1, bool=1, bool=0, bool=1, bool=0>
296.56us  void k_nonbonded_unified<float, bool=1, bool=0, bool=0, bool=1, bool=0>
256.41us  void k_nonbonded_unified<float, bool=1, bool=0, bool=1, bool=0, bool=0> <- du_dx only
247.96us  void k_nonbonded_unified<float, bool=1, bool=1, bool=0, bool=0, bool=1>
224.96us  void k_nonbonded_unified<float, bool=1, bool=0, bool=0, bool=0, bool=1>
170.47us  void k_nonbonded_unified<float, bool=1, bool=1, bool=0, bool=0, bool=0>
```

Our goal is to bring the performance of ```du_dx and du_dl``` to nearly that of ```du_dx``` only. Note that the ```alchem``` template parameter is an implementation detail.

Another nice side effect of this (hopefully) is that we also get a nice speed up from not having to compute host-host interactions in 3D (even for the du_dx only) case.