@maxentile 2021-02-19
This is great! The added flexibility of being able to toggle any of these combinations should also be a huge help down the road for other applications besides the ones you listed, as we need only some of these combinations for certain kinds of post-processing. For example, reweighting requires `u` only (or in some cases `u` and `du_dl` only), which already appears significantly sped up in the table you posted.