@proteneer 2021-02-19
```
547.11us  void k_nonbonded_unified<float, bool=1, bool=1, bool=1, bool=1>
524.21us  void k_nonbonded_unified<float, bool=0, bool=1, bool=1, bool=1>
520.10us  void k_nonbonded_unified<float, bool=1, bool=1, bool=0, bool=1>
490.51us  void k_nonbonded_unified<float, bool=0, bool=1, bool=0, bool=1>
398.44us  void k_nonbonded_unified<float, bool=1, bool=1, bool=1, bool=0>
379.36us  void k_nonbonded_unified<float, bool=1, bool=1, bool=0, bool=0>
309.35us  void k_nonbonded_unified<float, bool=1, bool=0, bool=1, bool=1>
298.79us  void k_nonbonded_unified<float, bool=0, bool=1, bool=1, bool=0> <-- du_dx and du_dl
298.50us  void k_nonbonded_unified<float, bool=1, bool=0, bool=0, bool=1>
290.65us  void k_nonbonded_unified<float, bool=0, bool=1, bool=0, bool=0> <-- du_dx only
274.72us  void k_nonbonded_unified<float, bool=0, bool=0, bool=1, bool=1>
268.50us  void k_nonbonded_unified<float, bool=0, bool=0, bool=0, bool=1>
189.92us  void k_nonbonded_unified<float, bool=1, bool=0, bool=1, bool=0>
186.81us  void k_nonbonded_unified<float, bool=1, bool=0, bool=0, bool=0>
```