@proteneer 2021-02-22
```
master
hif2a-apo: N=8805 speed: 512.80ns/day dt: 1.5fs (ran 25000 steps in 6.32s)
hif2a-rbfe-with-du-dp: N=8840 speed: 408.32ns/day dt: 1.5fs (ran 25000 steps in 7.94s)
hif2a-rbfe-du-dl-freq-0: N=8840 speed: 413.40ns/day dt: 1.5fs (ran 25000 steps in 7.84s)
hif2a-rbfe-du-dl-freq-1: N=8840 speed: 230.82ns/day dt: 1.5fs (ran 25000 steps in 14.04s)
hif2a-rbfe-du-dl-freq-5: N=8840 speed: 357.52ns/day dt: 1.5fs (ran 25000 steps in 9.06s)

this branch
hif2a-apo: N=8805 speed: 501.13ns/day dt: 1.5fs (ran 25000 steps in 6.47s)
hif2a-rbfe-with-du-dp: N=8840 speed: 416.04ns/day dt: 1.5fs (ran 25000 steps in 7.79s)
hif2a-rbfe-du-dl-freq-0: N=8840 speed: 416.81ns/day dt: 1.5fs (ran 25000 steps in 7.78s)
hif2a-rbfe-du-dl-freq-1: N=8840 speed: 396.17ns/day dt: 1.5fs (ran 25000 steps in 8.18s)
hif2a-rbfe-du-dl-freq-5: N=8840 speed: 412.09ns/day dt: 1.5fs (ran 25000 steps in 7.86s)
```
72% speed-up for hif2a-rbfe-du-dl-freq-1 (computing du/dl on every step), and the overall performance is within 5% of hif2a-rbfe-du-dl-freq-0 (only computing forces!)