@proteneer 2021-02-22
List of optimizations implemented in this PR
- Generate nonbonded kernels to compute selectively based on template args that elide unnecessary arguments
- Separates each tile into a fast "non-alchemical" tile, and a slow "alchemical" tile based on doing a fast check at the start of the calculation
- Modifies du_dl computation and du_dp computations to require slightly less floating point operations, in particular, an extra erfc was removed in the du_dl part by re-using a pre-computed number. (erfcs are one of the slowest operations we can do). 
- Modifies the kernel code to always accumulate du_dl and u, which are scalars, to instead a buffered array. The final array is reduced using cub in a single, final pass, as opposed to relying on atomic instructions.