@proteneer 2021-02-23
Discussed with @schmolly offline re: docking changes. There's a lot of perf we can squeeze out still re: re-using contexts and what not, but in the interest of keeping scope limited we'll do that later.